Update/Patch/Remove
Improve the performance:
Eliminate Gemini
Reason:
it reduces performance: we cannot force users to wait for gemini to answer the question, it takes a long time to answer, slowing down the website.
In addition, because the answer is long, having to cut it down makes it unnecessary, and we cannot lead to a conversation like Microsoft Bing.
Finally, Gemini belongs to Google, so the fact that it uses Google to search makes it unnecessary for Neutron to have it
Improved links in the interface
Update contribution feature:
Remove Insert, Change, Remove, GetID because it is no longer necessary, instead, we use crawler to do the work, users will only need to enter the website url
Combine Manager, Search, App into one to improve both performance and hosting
Improved the homepage image to fit the size